refactor(organization-main): member, restaurant 관련 파일 이름 복수형으로 통일
