+++ src/feature/organization/organizationMembers/components/section/OrganizationMembersSection.tsx
@@ -4,7 +4,7 @@ import Button from "components/button/Button";
 import Member from "components/dataDisplay/Member";
 import Modal from "components/modal/Modal";
 import { useState, useEffect } from "react";
-import { TMember } from "feature/organization/organizationMember/types/TMember";
+import { TMember } from "feature/organization/organizationMembers/types/TMember";
 
 interface Props {
   organizationId: string | number | string[];
